Treat nearley and nearleyc as separate modules
This aids documentation, as well as properly separating the core parsing algorithm from the grammar specification language.
