+++ examples/calculator/calculator.js
@@ -1,6 +1,6 @@
 // This is an example of how to use a nearley-made grammar.
 var PROMPT = "> ";
-var nearley = require("../../lib/nearley.js");
+var nearley = require("../../nearley");
 var grammar = require("./grammar.js");
 
 // This is where the action is.
