+++ nearleyc/compile.js
@@ -1,6 +1,7 @@
 (function(root, factory) {
+    // TODO this is not proper UMD, there are bare require()s below
     if (typeof module === 'object' && module.exports) {
-        module.exports = factory(require('./nearley'));
+        module.exports = factory(require('../nearley'));
     } else {
         root.Compile = factory(root.nearley);
     }
@@ -46,11 +47,11 @@ function Compile(structure, opts) {
             if (opts.alreadycompiled.indexOf(path) === -1) {
                 opts.alreadycompiled.push(path);
                 f = require('fs').readFileSync(path).toString();
-                var parserGrammar = new require('./nearley-language-bootstrapped.js');
+                var parserGrammar = new require('./nearley-language-bootstrapped');
                 var parser = new nearley.Parser(parserGrammar.ParserRules, parserGrammar.ParserStart);
                 parser.feed(f);
                 var c = Compile(parser.results[0], {path: path, __proto__:opts});
-                require('./lint.js')(c, {out: process.stderr});
+                require('./lint')(c, {out: process.stderr});
                 result.rules = result.rules.concat(c.rules);
                 result.body  = result.body.concat(c.body);
                 result.customTokens = result.customTokens.concat(c.customTokens);
