+++ nearleyc/stream.js
